+++ openssl-sys/src/handwritten/types.rs
@@ -3,14 +3,27 @@ use libc::*;
 #[allow(unused_imports)]
 use super::super::*;
 
+pub enum ASN1_OBJECT {}
+pub enum ASN1_VALUE {}
+
+pub type ASN1_BOOLEAN = c_int;
 pub enum ASN1_INTEGER {}
+pub enum ASN1_ENUMERATED {}
 pub enum ASN1_GENERALIZEDTIME {}
 pub enum ASN1_STRING {}
 pub enum ASN1_BIT_STRING {}
 pub enum ASN1_TIME {}
-pub enum ASN1_TYPE {}
-pub enum ASN1_OBJECT {}
 pub enum ASN1_OCTET_STRING {}
+pub enum ASN1_NULL {}
+pub enum ASN1_PRINTABLESTRING {}
+pub enum ASN1_T61STRING {}
+pub enum ASN1_IA5STRING {}
+pub enum ASN1_GENERALSTRING {}
+pub enum ASN1_BMPSTRING {}
+pub enum ASN1_UNIVERSALSTRING {}
+pub enum ASN1_UTCTIME {}
+pub enum ASN1_VISIBLESTRING {}
+pub enum ASN1_UTF8STRING {}
 
 pub enum bio_st {} // FIXME remove
 cfg_if! {
@@ -111,6 +124,8 @@ pub enum PKCS8_PRIV_KEY_INFO {}
 pub enum EVP_PKEY_ASN1_METHOD {}
 
 pub enum EVP_PKEY_CTX {}
+
+pub enum CMAC_CTX {}
 
 cfg_if! {
     if #[cfg(any(ossl110, libressl280))] {
@@ -324,6 +339,8 @@ cfg_if! {
         }
     }
 }
+
+stack!(stack_st_X509_ALGOR);
 
 pub enum X509_LOOKUP_METHOD {}