+++ openssl/src/ssl/mod.rs
@@ -659,6 +659,17 @@ impl SslVersion {
 
     #[cfg(babassl)]
     pub const NTLS1_1: SslVersion = SslVersion(ffi::NTLS1_1_VERSION);
+
+    /// DTLSv1.0
+    ///
+    /// DTLS 1.0 corresponds to TLS 1.1.
+    pub const DTLS1: SslVersion = SslVersion(ffi::DTLS1_VERSION);
+
+    /// DTLSv1.2
+    ///
+    /// DTLS 1.2 corresponds to TLS 1.2 to harmonize versions. There was never a DTLS 1.1.
+    #[cfg(any(ossl102, libressl332))]
+    pub const DTLS1_2: SslVersion = SslVersion(ffi::DTLS1_2_VERSION);
 }
 
 cfg_if! {
@@ -1279,7 +1290,7 @@ impl SslContextBuilder {
 
     /// Sets the minimum supported protocol version.
     ///
-    /// A value of `None` will enable protocol versions down the the lowest version supported by
+    /// A value of `None` will enable protocol versions down to the lowest version supported by
     /// OpenSSL.
     ///
     /// Requires OpenSSL 1.1.0 or LibreSSL 2.6.1 or newer.
@@ -1297,7 +1308,7 @@ impl SslContextBuilder {
 
     /// Sets the maximum supported protocol version.
     ///
-    /// A value of `None` will enable protocol versions down the the highest version supported by
+    /// A value of `None` will enable protocol versions up to the highest version supported by
     /// OpenSSL.
     ///
     /// Requires OpenSSL 1.1.0 or or LibreSSL 2.6.1 or newer.
@@ -1315,7 +1326,7 @@ impl SslContextBuilder {
 
     /// Gets the minimum supported protocol version.
     ///
-    /// A value of `None` indicates that all versions down the the lowest version supported by
+    /// A value of `None` indicates that all versions down to the lowest version supported by
     /// OpenSSL are enabled.
     ///
     /// Requires OpenSSL 1.1.0g or LibreSSL 2.7.0 or newer.
@@ -1334,7 +1345,7 @@ impl SslContextBuilder {
 
     /// Gets the maximum supported protocol version.
     ///
-    /// A value of `None` indicates that all versions down the the highest version supported by
+    /// A value of `None` indicates that all versions up to the highest version supported by
     /// OpenSSL are enabled.
     ///
     /// Requires OpenSSL 1.1.0g or LibreSSL 2.7.0 or newer.
@@ -3486,7 +3497,7 @@ impl SslRef {
 
     /// Sets the minimum supported protocol version.
     ///
-    /// A value of `None` will enable protocol versions down the the lowest version supported by
+    /// A value of `None` will enable protocol versions down to the lowest version supported by
     /// OpenSSL.
     ///
     /// Requires OpenSSL 1.1.0 or LibreSSL 2.6.1 or newer.
@@ -3504,7 +3515,7 @@ impl SslRef {
 
     /// Sets the maximum supported protocol version.
     ///
-    /// A value of `None` will enable protocol versions down the the highest version supported by
+    /// A value of `None` will enable protocol versions up to the highest version supported by
     /// OpenSSL.
     ///
     /// Requires OpenSSL 1.1.0 or or LibreSSL 2.6.1 or newer.