+++ systest/build.rs
@@ -56,6 +56,7 @@ fn main() {
         .header("openssl/bio.h")
         .header("openssl/x509v3.h")
         .header("openssl/safestack.h")
+        .header("openssl/cmac.h")
         .header("openssl/hmac.h")
         .header("openssl/obj_mac.h")
         .header("openssl/ssl.h")
@@ -110,7 +111,11 @@ fn main() {
             || s.starts_with("CRYPTO_EX_")
     });
     cfg.skip_struct(|s| {
-        s == "ProbeResult" || s == "X509_OBJECT_data" // inline union
+        s == "ProbeResult" ||
+            s == "X509_OBJECT_data" || // inline union
+            s == "DIST_POINT_NAME_st_anon_union" || // inline union
+            s == "PKCS7_data" ||
+            s == "ASN1_TYPE_value"
     });
     cfg.skip_fn(move |s| {
         s == "CRYPTO_memcmp" ||                 // uses volatile
@@ -130,7 +135,10 @@ fn main() {
     cfg.skip_field_type(|s, field| {
         (s == "EVP_PKEY" && field == "pkey") ||      // union
             (s == "GENERAL_NAME" && field == "d") || // union
-            (s == "X509_OBJECT" && field == "data") // union
+            (s == "DIST_POINT_NAME" && field == "name") || // union
+            (s == "X509_OBJECT" && field == "data") || // union
+            (s == "PKCS7" && field == "d") || // union
+            (s == "ASN1_TYPE" && field == "value") // union
     });
     cfg.skip_signededness(|s| {
         s.ends_with("_cb")